# Speed-up metric maintenance for component method calls.